+++ CHANGELOG.md
@@ -1,5 +1,12 @@
+1.2.0
+- Added Compress & Extract (tar/zip)
+- Added "Save as Root" and "Save as Specified User..." buttons/icons directly to the Editor Actions Title bar for quick access
+- Added `remote-ssh-sudo.timeoutSeconds` setting to allow users to increase timeout for slow network operations
+- Multi-Selection Delete Permanently as Root
+- Improved Security
+
 1.1.0
-- Add chmod & chown
+- Added chmod & chown
 
 1.0.0
 - Initial release